Python/EmailAPI/requirements.txt
grpcio==1.41.0
grpcio-tools==1.41.0
